Fix mkstemp usage and descriptor leak.
This code didn't work from the beginning!
mkstemp requires 6 X's, not 5.
Compating local_name to NULL doesn't work because local_name is a const label for an array in stack.
It can't be NULL.
If mkstemp fails, it returns -1. If not, it returns an descriptor to open file. We need to close it
to prevent descriptor from leaking. 
Well, program is terminated a couple of lines below, so the leak isn't fatal, but I just don't get this
as an excuse.



git-svn-id: file:///.data/svn_convert/svn@7577 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d